+++ demo/euphoria_showcase.gd
@@ -4,6 +4,8 @@
 ## Right: Injuries (regional impairment, micro reactions, threat anticipation)
 extends Node3D
 
+const DemoHelpers := preload("res://demo/demo_helpers.gd")
+
 const WALK_SPEED := 1.8
 
 var _profiles: Array[ImpactProfile] = []
@@ -70,16 +72,12 @@ func _ready() -> void:
 		moving_ctrl.stagger_finished.connect(_on_mover_stagger_finished)
 
 	# Start walk animation
-	_moving_anim = _find_child_of_type(_char_roots[1], "AnimationPlayer")
+	_moving_anim = DemoHelpers.find_descendant_of_type(_char_roots[1], "AnimationPlayer")
 	if _moving_anim:
 		_moving_anim.call_deferred("play", "walk")
 
 	# Debug gizmos
-	var debug_hud := StrengthDebugHUD.new()
-	debug_hud.name = "StrengthDebugHUD"
-	debug_hud.mouse_filter = Control.MOUSE_FILTER_IGNORE
-	debug_hud.set_anchors_and_offsets_preset(Control.PRESET_FULL_RECT)
-	add_child(debug_hud)
+	DemoHelpers.add_debug_hud(self)
 
 
 # --- Moving target signal handlers ---
@@ -104,62 +102,10 @@ func _on_mover_stagger_finished() -> void:
 # --- Setup ---
 
 func _setup_active(char_root: Node3D) -> KickbackCharacter:
-	var ybot_name := ""
-	for child in char_root.get_children():
-		if _find_child_of_type(child, "Skeleton3D"):
-			ybot_name = child.name
-			break
-	if ybot_name.is_empty():
-		return null
-
-	var skel_path := NodePath("../%s/Skeleton3D" % ybot_name)
-	var root_path := NodePath("..")
-	var builder_path := NodePath("../PhysicsRigBuilder")
-	var spring_path := NodePath("../SpringResolver")
-
-	var rb := PhysicsRigBuilder.new()
-	rb.name = "PhysicsRigBuilder"
-	rb.skeleton_path = skel_path
-	var rs := PhysicsRigSync.new()
-	rs.name = "PhysicsRigSync"
-	rs.skeleton_path = skel_path
-	rs.rig_builder_path = builder_path
-	var sp := SpringResolver.new()
-	sp.name = "SpringResolver"
-	sp.skeleton_path = skel_path
-	sp.rig_builder_path = builder_path
-	var ac := ActiveRagdollController.new()
-	ac.name = "ActiveRagdollController"
-	ac.spring_resolver_path = spring_path
-	ac.rig_builder_path = builder_path
-	ac.character_root_path = root_path
-	_controllers.append(ac)
-
-	var tuning := RagdollTuning.create_default()
-
-	var kc := KickbackCharacter.new()
-	kc.name = "KickbackCharacter"
-	kc.skeleton_path = skel_path
-	kc.character_root_path = root_path
-	kc.ragdoll_profile = RagdollProfile.create_mixamo_default()
-	kc.ragdoll_tuning = tuning
-
-	char_root.add_child(rb)
-	char_root.add_child(rs)
-	char_root.add_child(sp)
-	char_root.add_child(ac)
-	char_root.add_child(kc)
+	var kc := DemoHelpers.build_active_rig(char_root)
+	if kc:
+		_controllers.append(kc.get_active_controller())
 	return kc
-
-
-func _find_child_of_type(node: Node, type_name: String) -> Node:
-	for child in node.get_children():
-		if child.get_class() == type_name:
-			return child
-		var found := _find_child_of_type(child, type_name)
-		if found:
-			return found
-	return null
 
 
 func _physics_process(delta: float) -> void:
@@ -184,16 +130,7 @@ func _physics_process(delta: float) -> void:
 func _update_camera() -> void:
 	if not _cam:
 		return
-	var pivot := Vector3(0, 1.0, 0)
-	var yaw_rad := deg_to_rad(_cam_yaw)
-	var pitch_rad := deg_to_rad(_cam_pitch)
-	var offset := Vector3(
-		sin(yaw_rad) * cos(pitch_rad),
-		-sin(pitch_rad),
-		cos(yaw_rad) * cos(pitch_rad),
-	) * _cam_distance
-	_cam.global_position = pivot + offset
-	_cam.look_at(pivot)
+	DemoHelpers.orbit_camera(_cam, _cam_yaw, _cam_pitch, _cam_distance)
 
 
 func _unhandled_input(event: InputEvent) -> void: